// LinkERP/ClientApp/src/app/models/sys/lbs-sys-company.ts
import { BaseEntity } from "../base-entity";

export class LBSSYSCompany extends BaseEntity {
    CompanyCode: any;
    Name: any;
    TradingName : any;
    Laddr1: any;
    Laddr2: any;
    Laddr3 : any;
    Paddr1: any;
    Paddr2: any;
    Paddr3: any;
    Country: any;
    State: any;
    City: any;
    ZipCode: any;
    TaxNumber: any;
    SuperannuationID: any;
    DefaultPayrollRounding: any;
    DefaultCurrency : any;
    Logo: any;
    Phone: any;
    Fax : any;
    EmailAddress: any;  
}